model_service/requirements.txt
flask
flask-cors
torch
torchvision
Pillow
streamlit
requests
psycopg2
streamlit-drawable-canvas